Saving graph.

+++ synth_data/combine.py
@@ -25,10 +25,10 @@ f.write('%d %d %d\n' % (dims[0], dims[1], 256))
 for x in range(0, dims[0]):
     for y in range(0, dims[1]):
         for z in range(0, 256):
-            value = pix1[x, y] * exp(-float(z) / tau1) + pix2[x, y] * exp(-float(z) / tau2)
+          value = int(float(pix1[x, y]) * exp(-float(z) / tau1) + float(pix2[x, y]) * exp(-float(z) / tau2))
 
-            line = '%d %d %d %d \n' % (x, y, z, int(value))
+          line = '%d %d %d %d \n' % (x, y, z, int(value))
 
-            f.write(line)
+          f.write(line)
 
 f.close()
